#59 Analytics Phase 3: 4 dashboard templates (executive, marketing, operations, partner)
- Add ExecutiveDashboardView (shared) with title/subtitle/defaultPeriod
- Executive page uses view with defaults
- Add /admin/analytics/marketing and /admin/analytics/operations
- Design doc: Phase 3 section and status Done
- Delivery focus: #59 completed, next Ready item
